+++ crates/common_enums/src/connector_enums.rs
@@ -93,7 +93,7 @@ pub enum RoutableConnectors {
     Klarna,
     Mifinity,
     Mollie,
-    // Moneris,
+    Moneris,
     Multisafepay,
     Nexinets,
     Nexixpay,
@@ -233,7 +233,7 @@ pub enum Connector {
     Klarna,
     Mifinity,
     Mollie,
-    // Moneris,
+    Moneris,
     Multisafepay,
     Netcetera,
     Nexinets,
@@ -318,6 +318,7 @@ impl Connector {
                 | (Self::Deutschebank, _)
                 | (Self::Globalpay, _)
                 | (Self::Jpmorgan, _)
+                | (Self::Moneris, _)
                 | (Self::Paypal, _)
                 | (Self::Payu, _)
                 | (Self::Trustpay, PaymentMethod::BankRedirect)
@@ -384,7 +385,7 @@ impl Connector {
             | Self::Klarna
             | Self::Mifinity
             | Self::Mollie
-            // | Self::Moneris
+            | Self::Moneris
             | Self::Multisafepay
             | Self::Nexinets
             | Self::Nexixpay
@@ -513,6 +514,7 @@ impl From<RoutableConnectors> for Connector {
             RoutableConnectors::Klarna => Self::Klarna,
             RoutableConnectors::Mifinity => Self::Mifinity,
             RoutableConnectors::Mollie => Self::Mollie,
+            RoutableConnectors::Moneris => Self::Moneris,
             RoutableConnectors::Multisafepay => Self::Multisafepay,
             RoutableConnectors::Nexinets => Self::Nexinets,
             RoutableConnectors::Nexixpay => Self::Nexixpay,