+++ crates/router/src/types/storage/payment_attempt.rs
@@ -218,6 +218,9 @@ mod tests {
             profile_id: common_utils::generate_profile_id_of_default_length(),
             organization_id: Default::default(),
             connector_mandate_detail: Default::default(),
+            request_extended_authorization: Default::default(),
+            extended_authorization_applied: Default::default(),
+            capture_before: Default::default(),
             card_discovery: Default::default(),
         };
 
@@ -302,6 +305,9 @@ mod tests {
             profile_id: common_utils::generate_profile_id_of_default_length(),
             organization_id: Default::default(),
             connector_mandate_detail: Default::default(),
+            request_extended_authorization: Default::default(),
+            extended_authorization_applied: Default::default(),
+            capture_before: Default::default(),
             card_discovery: Default::default(),
         };
         let store = state
@@ -399,6 +405,9 @@ mod tests {
             profile_id: common_utils::generate_profile_id_of_default_length(),
             organization_id: Default::default(),
             connector_mandate_detail: Default::default(),
+            request_extended_authorization: Default::default(),
+            extended_authorization_applied: Default::default(),
+            capture_before: Default::default(),
             card_discovery: Default::default(),
         };
         let store = state